Added a simple file caching gizmo so that packfiles are only opened once when drilling down through delta objects. (This fixes the problem where have ~50 instances of the pack file open caused the OS to not allow any more files to be opened.)
There still remains a problem where for each delta object opened, a tmpfile is created; to build one full object that is deltafied, this can mean over 50 tmpfiles at a time (which again, the OS likely won't be happy about). This needs to be addressed by using memory instead of tmpfiles.